🚸 Integrate <DictEntry> into chapter examples
Note about the author's experience:

The reason why <DictEntry> takes an id rather than a name is because the
id is expected to be immutable. Eberban is very much still in progress
and even a word's name is subject to change.

I leave comments for the intended name I want so that future
contributors may understand what examples I wanted to display.