📝 Update Word Types and morphology to v0.56
While updating to v0.56 I had the idea to use a terse
(yet understandable notation) for morphology. The English instructions
were getting wordy and I wanted something pithy to summarise them. The
symbols are introduced gradually so as not to overwhelm.

Ideally, this notation is so good that people in the community start
using it.

And yes, it's inspired by RegEx.

This idea led to these smaller commits.

💄 Add "Fira Mono" for monospace font
✨ Add Symbolic component

And I also decided to tidy up the wording and structure of the whole
chapter, including this commit.

💄 Update Word Types component to match chapter structure